feat(k8s): update add task_strategy abstraction to custom strategy impl
