+++ kubernetes/internal/controller/batchsandbox_controller_test.go
@@ -51,7 +51,6 @@ import (
 	taskscheduler "github.com/alibaba/OpenSandbox/sandbox-k8s/internal/scheduler"
 	mock_scheduler "github.com/alibaba/OpenSandbox/sandbox-k8s/internal/scheduler/mock"
 	"github.com/alibaba/OpenSandbox/sandbox-k8s/internal/utils/fieldindex"
-	api "github.com/alibaba/OpenSandbox/sandbox-k8s/pkg/task-executor"
 )
 
 func init() {
@@ -696,158 +695,6 @@ func TestBatchSandboxReconciler_scheduleTasks(t *testing.T) {
 	}
 }
 
-func Test_getTaskSpec(t *testing.T) {
-	type args struct {
-		batchSbx *sandboxv1alpha1.BatchSandbox
-		idx      int
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    *api.Task
-		wantErr bool
-	}{
-		{
-			name: "basic task spec without patches",
-			args: args{
-				batchSbx: &sandboxv1alpha1.BatchSandbox{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "test-bs",
-						Namespace: "default",
-					},
-					Spec: sandboxv1alpha1.BatchSandboxSpec{
-						TaskTemplate: &sandboxv1alpha1.TaskTemplateSpec{
-							Spec: sandboxv1alpha1.TaskSpec{
-								Process: &sandboxv1alpha1.ProcessTask{
-									Command: []string{"echo", "hello"},
-								},
-							},
-						},
-					},
-				},
-				idx: 0,
-			},
-			want: &api.Task{
-				Name: "test-bs-0",
-				Process: &api.Process{
-					Command: []string{"echo", "hello"},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "task spec with shard patch",
-			args: args{
-				batchSbx: &sandboxv1alpha1.BatchSandbox{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "test-bs",
-						Namespace: "default",
-					},
-					Spec: sandboxv1alpha1.BatchSandboxSpec{
-						TaskTemplate: &sandboxv1alpha1.TaskTemplateSpec{
-							Spec: sandboxv1alpha1.TaskSpec{
-								Process: &sandboxv1alpha1.ProcessTask{
-									Command: []string{"echo", "hello"},
-								},
-							},
-						},
-						ShardTaskPatches: []runtime.RawExtension{
-							{
-								Raw: []byte(`{"spec":{"process":{"command":["echo","world"]}}}`),
-							},
-						},
-					},
-				},
-				idx: 0,
-			},
-			want: &api.Task{
-				Name: "test-bs-0",
-				Process: &api.Process{
-					Command: []string{"echo", "world"},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "task spec with invalid patch",
-			args: args{
-				batchSbx: &sandboxv1alpha1.BatchSandbox{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "test-bs",
-						Namespace: "default",
-					},
-					Spec: sandboxv1alpha1.BatchSandboxSpec{
-						TaskTemplate: &sandboxv1alpha1.TaskTemplateSpec{
-							Spec: sandboxv1alpha1.TaskSpec{
-								Process: &sandboxv1alpha1.ProcessTask{
-									Command: []string{"echo", "hello"},
-								},
-							},
-						},
-						ShardTaskPatches: []runtime.RawExtension{
-							{
-								Raw: []byte(`{"invalid json`),
-							},
-						},
-					},
-				},
-				idx: 0,
-			},
-			want:    nil,
-			wantErr: true,
-		},
-		{
-			name: "task spec with index out of range patch",
-			args: args{
-				batchSbx: &sandboxv1alpha1.BatchSandbox{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "test-bs",
-						Namespace: "default",
-					},
-					Spec: sandboxv1alpha1.BatchSandboxSpec{
-						TaskTemplate: &sandboxv1alpha1.TaskTemplateSpec{
-							Spec: sandboxv1alpha1.TaskSpec{
-								Process: &sandboxv1alpha1.ProcessTask{
-									Command: []string{"echo", "hello"},
-								},
-							},
-						},
-						ShardTaskPatches: []runtime.RawExtension{
-							{
-								Raw: []byte(`{"spec":{"process":{"command":["echo","world"]}}}`),
-							},
-						},
-					},
-				},
-				idx: 1,
-			},
-			want: &api.Task{
-				Name: "test-bs-1",
-				Process: &api.Process{
-					Command: []string{"echo", "hello"},
-				},
-			},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := getTaskSpec(tt.args.batchSbx, tt.args.idx)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("getTaskSpec() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !tt.wantErr {
-				if got.Name != tt.want.Name {
-					t.Errorf("getTaskSpec() name = %v, want %v", got.Name, tt.want.Name)
-				}
-				if !reflect.DeepEqual(got.Process, tt.want.Process) {
-					t.Errorf("getTaskSpec() spec = %v, want %v", got.Process, tt.want.Process)
-				}
-			}
-		})
-	}
-}
 
 func Test_parseIndex(t *testing.T) {
 	type args struct {
